requirements.txt
asgiref==3.3.1
astpretty==2.1.0
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
cognitive-complexity==1.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
dateparser==1.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
eradicate==2.0.0
flake8==3.8.4
flake8-blind-except==0.2.0
flake8-builtins==1.5.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.0.0
flake8-comprehensions==3.3.1
flake8-django==1.1.1
flake8-eradicate==1.0.0
flake8-expression-complexity==0.0.9
flake8-functions==0.0.5
flake8-logging-format==0.6.0
flake8-polyfill==1.0.2
flake8-print==4.0.0
flake8-variables-names==0.0.4
humanize==3.2.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
maya==0.6.1
mccabe==0.6.1
mr-proper==0.0.6
packaging==20.9
pendulum==2.1.2
pep8-naming==0.11.1
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
pytzdata==2020.1
regex==2020.11.13
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
snaptime==0.2.4
sqlparse==0.4.1
stdlib-list==0.8.0
toml==0.10.2
tzlocal==2.1
unify==0.5
untokenize==0.1.1
uritemplate==3.0.1
urllib3==1.26.3
